+++ frontend/src/routes/FunscriptPage.tsx
