+++ frontend/src/routes/StashConfig.tsx
