+++ frontend/src/routes/clips/ClipPreviewPage.tsx
