+++ frontend/src/routes/library/videos.tsx
@@ -88,11 +88,14 @@ export default function ListVideos() {
     <>
       <JumpToTop />
 
-      <EditVideoModal
-        onClose={() => setEditingVideo(undefined)}
-        isOpen={Boolean(editingVideo)}
-        videoId={editingVideo}
-      />
+      {editingVideo && (
+        <EditVideoModal
+          onClose={() => setEditingVideo(undefined)}
+          isOpen={Boolean(editingVideo)}
+          videoId={editingVideo}
+        />
+      )}
+
       <div className="my-4 grid grid-cols-3 items-center">
         <div className="flex gap-2">
           <Link to="add" className="btn btn-accent">
